+++ packages/core/src/striper/Striper.tsx
@@ -1,22 +1,21 @@
-import React, {ReactNode, Children, FC} from 'react'
+import React, {ReactNode, FC, Children, Fragment} from 'react'
 
 import {Card} from '../primitive'
-import {Block} from './Block'
 
-export const LinedBlock: FC = ({children}) => {
+export const Striper: FC = ({children}) => {
   const elements = Children.toArray(children).filter(child => !!child)
   if (elements.length === 0) {
     return null
   }
   return (
-    <Block>
+    <Fragment>
       {Children.map(elements, (child: ReactNode, key: number) => (
         <Card
           key={key}
+          s={key === 0 ? undefined : '0 -1px 0 0 #e6e6e6'}
           children={child}
-          bt={key === 0 ? undefined : '1px solid #eee'}
         />
       ))}
-    </Block>
+    </Fragment>
   )
 }